trainer/requirements.txt
prometheus_client
mlflow==2.9.2
pandas
numpy
python-dotenv